+++ src/components/AICard.jsx
