+++ src/components/GadgetCard.jsx
