--- src/components/header.jsx
@@ -1,8 +1,0 @@
-export const Header = () => {
-  return (
-    <div className="flex justify-end border border-x-0 bg-white px-4 py-4 sm:px-6">
-      <h1 className="text-xl font-semibold">Admin</h1>
-      <img className="rounded-full"></img>
-    </div>
-  );
-};
